Start splitting up LspStdio into smaller modules
