+++ haskell-ide-engine.cabal
@@ -21,6 +21,9 @@ library
   hs-source-dirs:      src
   exposed-modules:     Haskell.Ide.Engine.Plugin.Base
                        Haskell.Ide.Engine.Dispatcher
+                       Haskell.Ide.Engine.LSP.CodeActions
+                       Haskell.Ide.Engine.LSP.Config
+                       Haskell.Ide.Engine.LSP.Reactor
                        Haskell.Ide.Engine.Options
                        Haskell.Ide.Engine.Plugin.ApplyRefact
                        Haskell.Ide.Engine.Plugin.Brittany
