+++ src/Haskell/Ide/Engine/Types.hs
@@ -2,7 +2,6 @@
 {-# LANGUAGE GADTs #-}
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE PatternSynonyms #-}
-{-# LANGUAGE OverloadedStrings #-}
 module Haskell.Ide.Engine.Types where
 
 import           Haskell.Ide.Engine.MonadTypes
